requirements.txt
flask>=3.0
flask-sock>=0.7
python-dotenv>=1.0
requests>=2.31
websockets>=12.0
